feat sixth flex exercise
This has been about laying out a whole home page
with an input field and buttons, header, main and
footer. I had trouble getting rid of the
text-decoration property of the links, but the
problem was that I was trying to get rid of the
underline of the links by styling the li element,
and I had to style the a tag element.